LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/list-interpolation-expected.txt


PASS CSS Transitions: property <transform> from [none] to [none] at (0.25) should be [none]
PASS CSS Transitions with transition: all: property <transform> from [none] to [none] at (0.25) should be [none]
FAIL CSS Animations: property <transform> from [none] to [none] at (0.25) should be [none] assert_equals: expected "none " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Web Animations: property <transform> from [none] to [none] at (0.25) should be [none] assert_equals: expected "none " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 0 ) "
PASS CSS Transitions: property <transform> from [none] to [translate(200px) rotate(720deg)] at (0.25) should be [translate(50px) rotate(180deg)]
PASS CSS Transitions with transition: all: property <transform> from [none] to [translate(200px) rotate(720deg)] at (0.25) should be [translate(50px) rotate(180deg)]
PASS CSS Animations: property <transform> from [none] to [translate(200px) rotate(720deg)] at (0.25) should be [translate(50px) rotate(180deg)]
PASS Web Animations: property <transform> from [none] to [translate(200px) rotate(720deg)] at (0.25) should be [translate(50px) rotate(180deg)]
PASS CSS Transitions: property <transform> from [translate(200px) rotate(720deg)] to [none] at (0.25) should be [translate(150px) rotate(540deg)]
PASS CSS Transitions with transition: all: property <transform> from [translate(200px) rotate(720deg)] to [none] at (0.25) should be [translate(150px) rotate(540deg)]
PASS CSS Animations: property <transform> from [translate(200px) rotate(720deg)] to [none] at (0.25) should be [translate(150px) rotate(540deg)]
PASS Web Animations: property <transform> from [translate(200px) rotate(720deg)] to [none] at (0.25) should be [translate(150px) rotate(540deg)]
FAIL CSS Transitions: property <transform> from [translate(100px)] to [translate(200px) rotate(720deg)] at (0.25) should be [translate(125px) rotate(180deg)] assert_equals: expected "matrix ( - 1 , 0 , 0 , - 1 , 125 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
FAIL CSS Transitions with transition: all: property <transform> from [translate(100px)] to [translate(200px) rotate(720deg)] at (0.25) should be [translate(125px) rotate(180deg)] assert_equals: expected "matrix ( - 1 , 0 , 0 , - 1 , 125 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
FAIL CSS Animations: property <transform> from [translate(100px)] to [translate(200px) rotate(720deg)] at (0.25) should be [translate(125px) rotate(180deg)] assert_equals: expected "matrix ( - 1 , 0 , 0 , - 1 , 125 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
FAIL Web Animations: property <transform> from [translate(100px)] to [translate(200px) rotate(720deg)] at (0.25) should be [translate(125px) rotate(180deg)] assert_equals: expected "matrix ( - 1 , 0 , 0 , - 1 , 125 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
FAIL CSS Transitions: property <transform> from [translate(100px) rotate(720deg)] to [translate(200px)] at (0.25) should be [translate(125px) rotate(540deg)] assert_equals: expected "matrix ( - 1 , 0 , 0 , - 1 , 125 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
FAIL CSS Transitions with transition: all: property <transform> from [translate(100px) rotate(720deg)] to [translate(200px)] at (0.25) should be [translate(125px) rotate(540deg)] assert_equals: expected "matrix ( - 1 , 0 , 0 , - 1 , 125 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
FAIL CSS Animations: property <transform> from [translate(100px) rotate(720deg)] to [translate(200px)] at (0.25) should be [translate(125px) rotate(540deg)] assert_equals: expected "matrix ( - 1 , 0 , 0 , - 1 , 125 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
FAIL Web Animations: property <transform> from [translate(100px) rotate(720deg)] to [translate(200px)] at (0.25) should be [translate(125px) rotate(540deg)] assert_equals: expected "matrix ( - 1 , 0 , 0 , - 1 , 125 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
PASS CSS Transitions: property <transform> from [scale(2) rotate(360deg) translate(100px) matrix(1, 0, 0, 1, 100, 0) skew(0deg)] to [scale(3) rotate(1080deg) translate(200px) matrix(1, 0, 0, 1, 0, 200) skew(720deg)] at (0.25) should be [scale(2.25) rotate(540deg) translate(125px) matrix(1, 0, 0, 1, 75, 50) skew(180deg)]
PASS CSS Transitions with transition: all: property <transform> from [scale(2) rotate(360deg) translate(100px) matrix(1, 0, 0, 1, 100, 0) skew(0deg)] to [scale(3) rotate(1080deg) translate(200px) matrix(1, 0, 0, 1, 0, 200) skew(720deg)] at (0.25) should be [scale(2.25) rotate(540deg) translate(125px) matrix(1, 0, 0, 1, 75, 50) skew(180deg)]
PASS CSS Animations: property <transform> from [scale(2) rotate(360deg) translate(100px) matrix(1, 0, 0, 1, 100, 0) skew(0deg)] to [scale(3) rotate(1080deg) translate(200px) matrix(1, 0, 0, 1, 0, 200) skew(720deg)] at (0.25) should be [scale(2.25) rotate(540deg) translate(125px) matrix(1, 0, 0, 1, 75, 50) skew(180deg)]
PASS Web Animations: property <transform> from [scale(2) rotate(360deg) translate(100px) matrix(1, 0, 0, 1, 100, 0) skew(0deg)] to [scale(3) rotate(1080deg) translate(200px) matrix(1, 0, 0, 1, 0, 200) skew(720deg)] at (0.25) should be [scale(2.25) rotate(540deg) translate(125px) matrix(1, 0, 0, 1, 75, 50) skew(180deg)]
FAIL CSS Transitions: property <transform> from [translateX(100px) scaleX(3) translate(500px) scale(2)] to [translateY(200px) scale(5) translateX(100px) scaleY(3)] at (0.25) should be [translate(75px, 50px) scale(3.5, 2) translate(400px, 0px) scale(1.75, 2.25)] assert_equals: expected "matrix ( 6.13 , 0 , 0 , 4.5 , 1475 , 50 ) " but got "matrix ( 5.75 , 0 , 0 , 5.25 , 1325 , 50 ) "
FAIL CSS Transitions with transition: all: property <transform> from [translateX(100px) scaleX(3) translate(500px) scale(2)] to [translateY(200px) scale(5) translateX(100px) scaleY(3)] at (0.25) should be [translate(75px, 50px) scale(3.5, 2) translate(400px, 0px) scale(1.75, 2.25)] assert_equals: expected "matrix ( 6.13 , 0 , 0 , 4.5 , 1475 , 50 ) " but got "matrix ( 5.75 , 0 , 0 , 5.25 , 1325 , 50 ) "
FAIL CSS Animations: property <transform> from [translateX(100px) scaleX(3) translate(500px) scale(2)] to [translateY(200px) scale(5) translateX(100px) scaleY(3)] at (0.25) should be [translate(75px, 50px) scale(3.5, 2) translate(400px, 0px) scale(1.75, 2.25)] assert_equals: expected "matrix ( 6.13 , 0 , 0 , 4.5 , 1475 , 50 ) " but got "matrix ( 5.75 , 0 , 0 , 5.25 , 1325 , 50 ) "
FAIL Web Animations: property <transform> from [translateX(100px) scaleX(3) translate(500px) scale(2)] to [translateY(200px) scale(5) translateX(100px) scaleY(3)] at (0.25) should be [translate(75px, 50px) scale(3.5, 2) translate(400px, 0px) scale(1.75, 2.25)] assert_equals: expected "matrix ( 6.13 , 0 , 0 , 4.5 , 1475 , 50 ) " but got "matrix ( 5.75 , 0 , 0 , 5.25 , 1325 , 50 ) "
FAIL CSS Transitions: property <transform> from [rotateX(90deg) translateX(100px)] to [rotate3d(50, 0, 0, 180deg) translateY(200px)] at (0.25) should be [rotateX(112.5deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , - 0.38 , 0.92 , 0 , 0 , - 0.92 , - 0.38 , 0 , 75 , - 19.13 , 46.19 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , - 0.38 , 0.92 , 0 , 0 , - 0.92 , - 0.38 , 0 , 75 , - 50 , 0 , 1 ) "
FAIL CSS Transitions with transition: all: property <transform> from [rotateX(90deg) translateX(100px)] to [rotate3d(50, 0, 0, 180deg) translateY(200px)] at (0.25) should be [rotateX(112.5deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , - 0.38 , 0.92 , 0 , 0 , - 0.92 , - 0.38 , 0 , 75 , - 19.13 , 46.19 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , - 0.38 , 0.92 , 0 , 0 , - 0.92 , - 0.38 , 0 , 75 , - 50 , 0 , 1 ) "
FAIL CSS Animations: property <transform> from [rotateX(90deg) translateX(100px)] to [rotate3d(50, 0, 0, 180deg) translateY(200px)] at (0.25) should be [rotateX(112.5deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , - 0.38 , 0.92 , 0 , 0 , - 0.92 , - 0.38 , 0 , 75 , - 19.13 , 46.19 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , - 0.38 , 0.92 , 0 , 0 , - 0.92 , - 0.38 , 0 , 75 , - 50 , 0 , 1 ) "
FAIL Web Animations: property <transform> from [rotateX(90deg) translateX(100px)] to [rotate3d(50, 0, 0, 180deg) translateY(200px)] at (0.25) should be [rotateX(112.5deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , - 0.38 , 0.92 , 0 , 0 , - 0.92 , - 0.38 , 0 , 75 , - 19.13 , 46.19 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , - 0.38 , 0.92 , 0 , 0 , - 0.92 , - 0.38 , 0 , 75 , - 50 , 0 , 1 ) "
FAIL CSS Transitions: property <transform> from [rotateX(90deg) translateX(100px)] to [rotateY(0deg) translateY(200px)] at (0.25) should be [rotateX(67.5deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.38 , 0.92 , 0 , 0 , - 0.92 , 0.38 , 0 , 75 , 19.13 , 46.19 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.38 , 0.92 , 0 , 0 , - 0.92 , 0.38 , 0 , 75 , 50 , 0 , 1 ) "
FAIL CSS Transitions with transition: all: property <transform> from [rotateX(90deg) translateX(100px)] to [rotateY(0deg) translateY(200px)] at (0.25) should be [rotateX(67.5deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.38 , 0.92 , 0 , 0 , - 0.92 , 0.38 , 0 , 75 , 19.13 , 46.19 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.38 , 0.92 , 0 , 0 , - 0.92 , 0.38 , 0 , 75 , 50 , 0 , 1 ) "
FAIL CSS Animations: property <transform> from [rotateX(90deg) translateX(100px)] to [rotateY(0deg) translateY(200px)] at (0.25) should be [rotateX(67.5deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.38 , 0.92 , 0 , 0 , - 0.92 , 0.38 , 0 , 75 , 19.13 , 46.19 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.38 , 0.92 , 0 , 0 , - 0.92 , 0.38 , 0 , 75 , 50 , 0 , 1 ) "
FAIL Web Animations: property <transform> from [rotateX(90deg) translateX(100px)] to [rotateY(0deg) translateY(200px)] at (0.25) should be [rotateX(67.5deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.38 , 0.92 , 0 , 0 , - 0.92 , 0.38 , 0 , 75 , 19.13 , 46.19 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.38 , 0.92 , 0 , 0 , - 0.92 , 0.38 , 0 , 75 , 50 , 0 , 1 ) "
FAIL CSS Transitions: property <transform> from [rotate3d(1, 1, 1, -60deg) translateX(100px)] to [rotate3d(2, 2, 2, 60deg) translateY(200px)] at (0.25) should be [rotate3d(1, 1, 1, -30deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 0.91 , - 0.24 , 0.33 , 0 , 0.33 , 0.91 , - 0.24 , 0 , - 0.24 , 0.33 , 0.91 , 0 , 84.97 , 27.23 , 12.8 , 1 ) " but got "matrix3d ( 0.91 , - 0.24 , 0.33 , 0 , 0.33 , 0.91 , - 0.24 , 0 , - 0.24 , 0.33 , 0.91 , 0 , 33.33 , 8.33 , 83.33 , 1 ) "
FAIL CSS Transitions with transition: all: property <transform> from [rotate3d(1, 1, 1, -60deg) translateX(100px)] to [rotate3d(2, 2, 2, 60deg) translateY(200px)] at (0.25) should be [rotate3d(1, 1, 1, -30deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 0.91 , - 0.24 , 0.33 , 0 , 0.33 , 0.91 , - 0.24 , 0 , - 0.24 , 0.33 , 0.91 , 0 , 84.97 , 27.23 , 12.8 , 1 ) " but got "matrix3d ( 0.91 , - 0.24 , 0.33 , 0 , 0.33 , 0.91 , - 0.24 , 0 , - 0.24 , 0.33 , 0.91 , 0 , 33.33 , 8.33 , 83.33 , 1 ) "
FAIL CSS Animations: property <transform> from [rotate3d(1, 1, 1, -60deg) translateX(100px)] to [rotate3d(2, 2, 2, 60deg) translateY(200px)] at (0.25) should be [rotate3d(1, 1, 1, -30deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 0.91 , - 0.24 , 0.33 , 0 , 0.33 , 0.91 , - 0.24 , 0 , - 0.24 , 0.33 , 0.91 , 0 , 84.97 , 27.23 , 12.8 , 1 ) " but got "matrix3d ( 0.91 , - 0.24 , 0.33 , 0 , 0.33 , 0.91 , - 0.24 , 0 , - 0.24 , 0.33 , 0.91 , 0 , 33.33 , 8.33 , 83.33 , 1 ) "
FAIL Web Animations: property <transform> from [rotate3d(1, 1, 1, -60deg) translateX(100px)] to [rotate3d(2, 2, 2, 60deg) translateY(200px)] at (0.25) should be [rotate3d(1, 1, 1, -30deg) translate(75px, 50px)] assert_equals: expected "matrix3d ( 0.91 , - 0.24 , 0.33 , 0 , 0.33 , 0.91 , - 0.24 , 0 , - 0.24 , 0.33 , 0.91 , 0 , 84.97 , 27.23 , 12.8 , 1 ) " but got "matrix3d ( 0.91 , - 0.24 , 0.33 , 0 , 0.33 , 0.91 , - 0.24 , 0 , - 0.24 , 0.33 , 0.91 , 0 , 33.33 , 8.33 , 83.33 , 1 ) "
FAIL CSS Transitions: property <transform> from [rotate3d(1, 0, 0, 360deg) translateX(100px)] to [rotate3d(0, 1, 0, -720deg) translateY(200px)] at (0.25) should be [rotate3d(0, 0, 1, 0deg) translate(75px, 50px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 75 , 50 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 75 , 50 , 0 , 1 ) "
FAIL CSS Transitions with transition: all: property <transform> from [rotate3d(1, 0, 0, 360deg) translateX(100px)] to [rotate3d(0, 1, 0, -720deg) translateY(200px)] at (0.25) should be [rotate3d(0, 0, 1, 0deg) translate(75px, 50px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 75 , 50 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 75 , 50 , 0 , 1 ) "
FAIL CSS Animations: property <transform> from [rotate3d(1, 0, 0, 360deg) translateX(100px)] to [rotate3d(0, 1, 0, -720deg) translateY(200px)] at (0.25) should be [rotate3d(0, 0, 1, 0deg) translate(75px, 50px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 75 , 50 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 75 , 50 , 0 , 1 ) "
FAIL Web Animations: property <transform> from [rotate3d(1, 0, 0, 360deg) translateX(100px)] to [rotate3d(0, 1, 0, -720deg) translateY(200px)] at (0.25) should be [rotate3d(0, 0, 1, 0deg) translate(75px, 50px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 75 , 50 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 75 , 50 , 0 , 1 ) "
FAIL CSS Transitions: property <transform> from [rotate(0deg) translate(100px)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [rotate(180deg) matrix(1.25, 0, 0, 1.25, 175, 0)] assert_equals: expected "matrix ( - 1.25 , 0 , 0 , - 1.25 , - 175 , 0 ) " but got "matrix ( 1.25 , 0 , 0 , 1.25 , 175 , 0 ) "
FAIL CSS Transitions with transition: all: property <transform> from [rotate(0deg) translate(100px)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [rotate(180deg) matrix(1.25, 0, 0, 1.25, 175, 0)] assert_equals: expected "matrix ( - 1.25 , 0 , 0 , - 1.25 , - 175 , 0 ) " but got "matrix ( 1.25 , 0 , 0 , 1.25 , 175 , 0 ) "
FAIL CSS Animations: property <transform> from [rotate(0deg) translate(100px)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [rotate(180deg) matrix(1.25, 0, 0, 1.25, 175, 0)] assert_equals: expected "matrix ( - 1.25 , 0 , 0 , - 1.25 , - 175 , 0 ) " but got "matrix ( 1.25 , 0 , 0 , 1.25 , 175 , 0 ) "
FAIL Web Animations: property <transform> from [rotate(0deg) translate(100px)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [rotate(180deg) matrix(1.25, 0, 0, 1.25, 175, 0)] assert_equals: expected "matrix ( - 1.25 , 0 , 0 , - 1.25 , - 175 , 0 ) " but got "matrix ( 1.25 , 0 , 0 , 1.25 , 175 , 0 ) "
PASS CSS Transitions: property <transform> from [scale(2) rotate(0deg) translate(100px)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [matrix(2, 0, 0, 2, 250, 0)]
PASS CSS Transitions with transition: all: property <transform> from [scale(2) rotate(0deg) translate(100px)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [matrix(2, 0, 0, 2, 250, 0)]
PASS CSS Animations: property <transform> from [scale(2) rotate(0deg) translate(100px)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [matrix(2, 0, 0, 2, 250, 0)]
PASS Web Animations: property <transform> from [scale(2) rotate(0deg) translate(100px)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [matrix(2, 0, 0, 2, 250, 0)]
PASS CSS Transitions: property <transform> from [scale(2) rotate(0deg)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [matrix(2, 0, 0, 2, 100, 0)]
PASS CSS Transitions with transition: all: property <transform> from [scale(2) rotate(0deg)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [matrix(2, 0, 0, 2, 100, 0)]
PASS CSS Animations: property <transform> from [scale(2) rotate(0deg)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [matrix(2, 0, 0, 2, 100, 0)]
PASS Web Animations: property <transform> from [scale(2) rotate(0deg)] to [rotate(720deg) scale(2) translate(200px)] at (0.25) should be [matrix(2, 0, 0, 2, 100, 0)]
FAIL CSS Transitions: property <transform> from [rotate(0deg) scaleX(1)] to [rotate(720deg) translateX(0px) scaleX(2)] at (0.25) should be [rotate(180deg) matrix(1.25, 0, 0, 1, 0, 0)] assert_equals: expected "matrix ( - 1.25 , 0 , 0 , - 1 , 0 , 0 ) " but got "matrix ( 1.25 , 0 , 0 , 1 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <transform> from [rotate(0deg) scaleX(1)] to [rotate(720deg) translateX(0px) scaleX(2)] at (0.25) should be [rotate(180deg) matrix(1.25, 0, 0, 1, 0, 0)] assert_equals: expected "matrix ( - 1.25 , 0 , 0 , - 1 , 0 , 0 ) " but got "matrix ( 1.25 , 0 , 0 , 1 , 0 , 0 ) "
FAIL CSS Animations: property <transform> from [rotate(0deg) scaleX(1)] to [rotate(720deg) translateX(0px) scaleX(2)] at (0.25) should be [rotate(180deg) matrix(1.25, 0, 0, 1, 0, 0)] assert_equals: expected "matrix ( - 1.25 , 0 , 0 , - 1 , 0 , 0 ) " but got "matrix ( 1.25 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Web Animations: property <transform> from [rotate(0deg) scaleX(1)] to [rotate(720deg) translateX(0px) scaleX(2)] at (0.25) should be [rotate(180deg) matrix(1.25, 0, 0, 1, 0, 0)] assert_equals: expected "matrix ( - 1.25 , 0 , 0 , - 1 , 0 , 0 ) " but got "matrix ( 1.25 , 0 , 0 , 1 , 0 , 0 ) "
FAIL CSS Transitions: property <transform> from [rotate(720deg) translateX(0px) scaleX(2)] to [rotate(0deg) scaleX(1)] at (0.25) should be [rotate(540deg) matrix(1.75, 0, 0, 1, 0, 0)] assert_equals: expected "matrix ( - 1.75 , 0 , 0 , - 1 , 0 , 0 ) " but got "matrix ( 1.75 , 0 , 0 , 1 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <transform> from [rotate(720deg) translateX(0px) scaleX(2)] to [rotate(0deg) scaleX(1)] at (0.25) should be [rotate(540deg) matrix(1.75, 0, 0, 1, 0, 0)] assert_equals: expected "matrix ( - 1.75 , 0 , 0 , - 1 , 0 , 0 ) " but got "matrix ( 1.75 , 0 , 0 , 1 , 0 , 0 ) "
FAIL CSS Animations: property <transform> from [rotate(720deg) translateX(0px) scaleX(2)] to [rotate(0deg) scaleX(1)] at (0.25) should be [rotate(540deg) matrix(1.75, 0, 0, 1, 0, 0)] assert_equals: expected "matrix ( - 1.75 , 0 , 0 , - 1 , 0 , 0 ) " but got "matrix ( 1.75 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Web Animations: property <transform> from [rotate(720deg) translateX(0px) scaleX(2)] to [rotate(0deg) scaleX(1)] at (0.25) should be [rotate(540deg) matrix(1.75, 0, 0, 1, 0, 0)] assert_equals: expected "matrix ( - 1.75 , 0 , 0 , - 1 , 0 , 0 ) " but got "matrix ( 1.75 , 0 , 0 , 1 , 0 , 0 ) "
FAIL CSS Transitions: property <transform> from [scaleX(-3) scaleY(2)] to [scaleY(-3) translateX(0px) scaleX(2)] at (0.25) should be [scale(-2, 0) matrix(1.25, 0, 0, 1.75, 0, 0)] assert_equals: expected "matrix ( - 2.5 , 0 , 0 , 0 , 0 , 0 ) " but got "matrix ( - 1.94 , - 1.94 , - 1.59 , 1.59 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <transform> from [scaleX(-3) scaleY(2)] to [scaleY(-3) translateX(0px) scaleX(2)] at (0.25) should be [scale(-2, 0) matrix(1.25, 0, 0, 1.75, 0, 0)] assert_equals: expected "matrix ( - 2.5 , 0 , 0 , 0 , 0 , 0 ) " but got "matrix ( - 1.94 , - 1.94 , - 1.59 , 1.59 , 0 , 0 ) "
FAIL CSS Animations: property <transform> from [scaleX(-3) scaleY(2)] to [scaleY(-3) translateX(0px) scaleX(2)] at (0.25) should be [scale(-2, 0) matrix(1.25, 0, 0, 1.75, 0, 0)] assert_equals: expected "matrix ( - 2.5 , 0 , 0 , 0 , 0 , 0 ) " but got "matrix ( - 1.94 , - 1.94 , - 1.59 , 1.59 , 0 , 0 ) "
FAIL Web Animations: property <transform> from [scaleX(-3) scaleY(2)] to [scaleY(-3) translateX(0px) scaleX(2)] at (0.25) should be [scale(-2, 0) matrix(1.25, 0, 0, 1.75, 0, 0)] assert_equals: expected "matrix ( - 2.5 , 0 , 0 , 0 , 0 , 0 ) " but got "matrix ( - 1.94 , - 1.94 , - 1.59 , 1.59 , 0 , 0 ) "
FAIL CSS Transitions: property <transform> from [scaleY(-3) translateX(0px) scaleX(2)] to [scaleX(-3) scaleY(2)] at (0.25) should be [scale(0, -2) matrix(1.75, 0, 0, 1.25, 0, 0)] assert_equals: expected "matrix ( 0 , 0 , 0 , - 2.5 , 0 , 0 ) " but got "matrix ( 1.59 , 1.59 , 1.94 , - 1.94 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <transform> from [scaleY(-3) translateX(0px) scaleX(2)] to [scaleX(-3) scaleY(2)] at (0.25) should be [scale(0, -2) matrix(1.75, 0, 0, 1.25, 0, 0)] assert_equals: expected "matrix ( 0 , 0 , 0 , - 2.5 , 0 , 0 ) " but got "matrix ( 1.59 , 1.59 , 1.94 , - 1.94 , 0 , 0 ) "
FAIL CSS Animations: property <transform> from [scaleY(-3) translateX(0px) scaleX(2)] to [scaleX(-3) scaleY(2)] at (0.25) should be [scale(0, -2) matrix(1.75, 0, 0, 1.25, 0, 0)] assert_equals: expected "matrix ( 0 , 0 , 0 , - 2.5 , 0 , 0 ) " but got "matrix ( 1.59 , 1.59 , 1.94 , - 1.94 , 0 , 0 ) "
FAIL Web Animations: property <transform> from [scaleY(-3) translateX(0px) scaleX(2)] to [scaleX(-3) scaleY(2)] at (0.25) should be [scale(0, -2) matrix(1.75, 0, 0, 1.25, 0, 0)] assert_equals: expected "matrix ( 0 , 0 , 0 , - 2.5 , 0 , 0 ) " but got "matrix ( 1.59 , 1.59 , 1.94 , - 1.94 , 0 , 0 ) "
FAIL CSS Transitions: property <transform> from [scaleY(-3) translateX(0px)] to [scaleX(-3) scaleY(2)] at (0.25) should be [scale(0, -2) matrix(1, 0, 0, 1.25, 0, 0)] assert_equals: expected "matrix ( 0 , 0 , 0 , - 2.5 , 0 , 0 ) " but got "matrix ( 1.06 , 1.06 , 1.94 , - 1.94 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <transform> from [scaleY(-3) translateX(0px)] to [scaleX(-3) scaleY(2)] at (0.25) should be [scale(0, -2) matrix(1, 0, 0, 1.25, 0, 0)] assert_equals: expected "matrix ( 0 , 0 , 0 , - 2.5 , 0 , 0 ) " but got "matrix ( 1.06 , 1.06 , 1.94 , - 1.94 , 0 , 0 ) "
FAIL CSS Animations: property <transform> from [scaleY(-3) translateX(0px)] to [scaleX(-3) scaleY(2)] at (0.25) should be [scale(0, -2) matrix(1, 0, 0, 1.25, 0, 0)] assert_equals: expected "matrix ( 0 , 0 , 0 , - 2.5 , 0 , 0 ) " but got "matrix ( 1.06 , 1.06 , 1.94 , - 1.94 , 0 , 0 ) "
FAIL Web Animations: property <transform> from [scaleY(-3) translateX(0px)] to [scaleX(-3) scaleY(2)] at (0.25) should be [scale(0, -2) matrix(1, 0, 0, 1.25, 0, 0)] assert_equals: expected "matrix ( 0 , 0 , 0 , - 2.5 , 0 , 0 ) " but got "matrix ( 1.06 , 1.06 , 1.94 , - 1.94 , 0 , 0 ) "
